Initial commit
Another addition to this year's Xmas posterity dump. It's a bit dated
and might suffer as a result. There's a certain update to http-backend
coming down the pike that's a likely candidate for disruption. No idea
when that's slated to occur, however: have yet to unravel the mysteries
of the Git mailing list and development process.